doc_chat_backend/requirements.txt
fastapi
uvicorn
python-multipart
unstructured
pyjwt
passlib[bcrypt]
sqlalchemy
python-dotenv
python-jose[cryptography]
elasticsearch
langchain
langchain-community
langchain-elasticsearch
crewai
tenacity
litellm
pydantic
sentence-transformers
transformers
torch
